Add ability to select file to import
To date the only way to import a file is to hope the exported file
is in the correct location. To remove this requirement, this change
checks if there are any activities avaialble on the system which
can look in the file system. If there are, new options for
importing data will be avaialble.

In addition, the import/export activity's layout was updated,
adding more explanation.